// src/types.ts
export type TLocationCoordinates = {
  latitude: number;
  longitude: number;
}

export type TCity = {
  title: string;
  lat: number;
  lng: number;
  zoom: number;
}

export type THost = {
  name: string;
  avatarUrl: string;
  isPro: boolean;
}

export type TOffer = {
  id: string;
  title: string;
  type: string;
  price: number;
  city: {
    name: string;
    location: {
      latitude: number;
      longitude: number;
      zoom: number;
    };
  };
  location: TLocationCoordinates;
  isFavorite: boolean;
  isPremium: boolean;
  rating: number;
  description: string;
  bedrooms: number;
  goods: string[];
  host: THost;
  images: string[];
  maxAdults: number;
}